+++ src/context/Interface.ts
@@ -3,8 +3,9 @@ import { User } from 'firebase/auth';
 // State
 export interface State {
   loading: boolean;
-  currentUser: User | null;
+  authorizedUser: User | null;
   registerApp: boolean;
   openModalAddEditElements: { modal: boolean; isEdit: boolean };
   openDeletedModal: boolean;
+  selectedId: string;
 }