+++ src/context/Reducer.ts
@@ -10,10 +10,10 @@ export const appReducer = (state: State, action: Action): State => {
         ...state,
         loading: action.payload,
       };
-    case 'SET_CURRENT_USER':
+    case 'SET_AUTHORIZED_USER':
       return {
         ...state,
-        currentUser: action.payload,
+        authorizedUser: action.payload,
       };
     case 'SET_REGISTER_APP':
       return {
@@ -32,6 +32,11 @@ export const appReducer = (state: State, action: Action): State => {
       return {
         ...state,
         openDeletedModal: action.payload,
+      };
+    case 'SET_SELECTED_ID':
+      return {
+        ...state,
+        selectedId: action.payload,
       };
 
     default: {