+++ src/ui/AddEditElementsModal/AddEditElementsModal.tsx
@@ -13,25 +13,106 @@ import NewElementForm from '../../components/NewElementForm';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { newElementSchema } from '../../schema/newElementSchema';
 import { ElementFormValues } from '../../types/newElementFormTypes';
+import { collection, doc, getDoc, setDoc, updateDoc } from 'firebase/firestore';
+import { db } from '../../api/firebase';
+import { ExpectedAPIFormat, dataPayloadNewElement } from '../../api/utils/dataPayloadNewElement';
+import { useEffect } from 'react';
+import { expectedElementFormValues } from '../../api/utils/expectedFrontedData';
+import { useSnackbar } from 'notistack';
 
 export default function AddEditElementsModal(): JSX.Element {
   const {
-    state: { openModalAddEditElements },
+    state: { openModalAddEditElements, selectedId, authorizedUser },
     dispatch,
   } = useAppState();
+
+  const { enqueueSnackbar } = useSnackbar();
 
   const onClose = () =>
     dispatch({ type: 'SET_OPEN_MODAL_ADD_EDIT_ELEMENTS', payload: { modal: false, isEdit: false } });
 
-  const methods = useForm({
+  const methods = useForm<ElementFormValues>({
     resolver: yupResolver(newElementSchema),
   });
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const onSubmit = (data: ElementFormValues) => {
-    // To DO: Data to send
-    // console.log(data);
-    onClose();
+  useEffect(() => {
+    if (openModalAddEditElements.isEdit && selectedId) {
+      getDocumentById(selectedId);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [openModalAddEditElements.isEdit, selectedId]);
+
+  /**
+   * ADD NEW ELEMENT:
+   */
+  const addNewOnSubmit = (data: ElementFormValues) => {
+    // Convert data to the expected format
+    const expectedFormatData = dataPayloadNewElement(data);
+
+    // Use a custom ID based on the email
+    const emailAsId = expectedFormatData.customer.email;
+
+    // Create a reference to the document with a custom ID
+    const docRef = doc(db, 'customers', authorizedUser?.email as string, 'users', emailAsId);
+
+    // Use setDoc instead of addDoc to set the custom ID
+    setDoc(docRef, expectedFormatData)
+      .then(() => {
+        onClose();
+        enqueueSnackbar('The new item has been created!', { variant: 'success' });
+      })
+      .catch((error) => {
+        enqueueSnackbar(error, { variant: 'error' });
+      });
+  };
+
+  /**
+   * EDIT ELEMENT:
+   */
+  const getDocumentById = (id: string) => {
+    // Create a reference to the collection
+    const ref = collection(db, 'customers', authorizedUser?.email as string, 'users');
+
+    // Create a reference to the document by ID
+    const docRef = doc(ref, id);
+
+    // Get the document
+    getDoc(docRef)
+      .then((docSnap) => {
+        if (docSnap.exists()) {
+          // Extract data from the document
+          const dataFromDB = docSnap.data();
+          // Convert data to the form values format
+          const dataInEditForm = expectedElementFormValues(dataFromDB as ExpectedAPIFormat);
+          // Set values to form
+          methods.reset(dataInEditForm);
+        }
+      })
+      .catch((error) => {
+        enqueueSnackbar(error, { variant: 'error' });
+      });
+  };
+
+  const editOnSubmit = (data: ElementFormValues) => {
+    // Prepare the data in the expected format
+    const expectedFormatData = dataPayloadNewElement(data);
+
+    // Use the email as the document ID
+    const emailAsId = expectedFormatData.customer.email;
+
+    // Create a reference to the document with the custom ID
+    const docRef = doc(db, 'customers', authorizedUser?.email as string, 'users', emailAsId);
+
+    // Update the document
+    updateDoc(docRef, expectedFormatData)
+      .then(() => {
+        // Close the dialog and show a success message
+        onClose();
+        enqueueSnackbar('The item has been edited!', { variant: 'warning' });
+      })
+      .catch((error) => {
+        enqueueSnackbar(error, { variant: 'error' });
+      });
   };
 
   return (
@@ -72,13 +153,19 @@ export default function AddEditElementsModal(): JSX.Element {
           <Divider />
 
           <FormProvider {...methods}>
-            <form onSubmit={methods.handleSubmit(onSubmit)}>
+            <form
+              onSubmit={
+                openModalAddEditElements.isEdit
+                  ? methods.handleSubmit(editOnSubmit)
+                  : methods.handleSubmit(addNewOnSubmit)
+              }
+            >
               <DialogContent>
                 <NewElementForm />
               </DialogContent>
 
               <DialogActions>
-                <Button variant='solid' color='success' type='submit'>
+                <Button variant='solid' color={openModalAddEditElements.isEdit ? 'warning' : 'success'} type='submit'>
                   Submit
                 </Button>
                 <Button variant='outlined' color='neutral' onClick={onClose}>