+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.2.5](https://github.com/MarcinMG-web/react-dashboard/compare/v0.1.6...v0.2.5) (2024-07-12)
+
+### Features
+
+- add CRUD operation ([9dc18dc](https://github.com/MarcinMG-web/react-dashboard/commit/9dc18dcfd00ce999c2f7a5a3f3a3b1ea70b2c16a))
+- add new form to add or eddit elements ([#6](https://github.com/MarcinMG-web/react-dashboard/issues/6)) ([2381d44](https://github.com/MarcinMG-web/react-dashboard/commit/2381d446cd7a114ae03d428a2f1b96336f8aa749))
+- each user has separate collection ([1d41c8d](https://github.com/MarcinMG-web/react-dashboard/commit/1d41c8dbafdd6c32986c423cc5d1d92506ec6107))
+- prepare custom hook and add notification ([4b2c32e](https://github.com/MarcinMG-web/react-dashboard/commit/4b2c32e27a1100626368e6e598a16cc5cfac2cd5))
+
+### Bug Fixes
+
+- back with protect main app ([e309f87](https://github.com/MarcinMG-web/react-dashboard/commit/e309f87be1c61201e94b7a252ef1dab37cface78))
+- display id in table ([ff9bc26](https://github.com/MarcinMG-web/react-dashboard/commit/ff9bc26c2d71f15abd147d8e2c46964313586ef3))
+- folder name ([e3dbecf](https://github.com/MarcinMG-web/react-dashboard/commit/e3dbecf283ee878b3f4720abb1437f04d7571f36))
+- name of variables in hook ([fd1a17e](https://github.com/MarcinMG-web/react-dashboard/commit/fd1a17e58463080c20985eb4ceb12ac9ec4666bf))
+- small bug in select ([c02136c](https://github.com/MarcinMG-web/react-dashboard/commit/c02136c78e7ca74bb58cee93c840449bd0d12aba))
+- small issue in code ([18f173f](https://github.com/MarcinMG-web/react-dashboard/commit/18f173f62c1f592bcf50732e6b839b65081f7be9))
+- udated style snackbar ([53416df](https://github.com/MarcinMG-web/react-dashboard/commit/53416df3f9cd4388ddcc94183ff3945de98f7119))
+
 ### 0.1.5 (2024-07-08)
 
 ### Features